make check work, avoid warnings
git-svn-id: https://dune.mathematik.uni-freiburg.de/svn/dune-fem/trunk@7306 4a6c9bf5-c326-0410-8cc6-c7cabc8ddcf0

+++ dune/fem/operator/1order/localmassmatrix.hh
@@ -382,9 +382,9 @@ protected:
                  const Rhs& rhs, 
                  X& x ) const 
   {
-    assert( matrix.rows() == size );
-    assert( matrix.cols() == size );
-    assert( rhs.size() == size );
+    assert( (int) matrix.rows() == size );
+    assert( (int) matrix.cols() == size );
+    assert( (int) rhs.size() == size );
 
     for( int row = 0; row < size; ++ row ) 
     {
